changed price type to double, added controller

+++ src/main/java/com/unimol/spring_backend/product/Product.java
@@ -19,5 +19,5 @@ public class Product {
     private String name;
     private String description;
     private String imgUrl;
-    private float price;
+    private double price;
 }
